Add Tcl stubs for noop executor
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@13116 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/turbine/async_exec.c
@@ -67,7 +67,7 @@ static bool executors_init = false;
 static struct table executors;
 
 static turbine_exec_code
-get_tasks(turbine_executor *executor, void *buffer, int buffer_size,
+get_tasks(turbine_executor *executor, void *buffer, size_t buffer_size,
           bool poll, int max_tasks);
 
 static turbine_exec_code
@@ -117,7 +117,8 @@ turbine_add_async_exec(turbine_executor executor)
 
 
 turbine_code
-turbine_async_worker_loop(const char *exec_name, void *buffer, int buffer_size)
+turbine_async_worker_loop(const char *exec_name, void *buffer,
+                          size_t buffer_size)
 {
   turbine_exec_code ec;
 
@@ -181,7 +182,7 @@ turbine_async_worker_loop(const char *exec_name, void *buffer, int buffer_size)
  * TODO: currently only executes one task, but could do multiple
  */
 static turbine_exec_code
-get_tasks(turbine_executor *executor, void *buffer, int buffer_size,
+get_tasks(turbine_executor *executor, void *buffer, size_t buffer_size,
           bool poll, int max_tasks)
 {
   adlb_code ac;
